+++ server/routes/media.ts
@@ -18,7 +18,7 @@ import logger from '@server/logger';
 import { isAuthenticated } from '@server/middleware/auth';
 import { Router } from 'express';
 import type { FindOneOptions } from 'typeorm';
-import { In } from 'typeorm';
+import { EntityNotFoundError, In, IsNull, Not } from 'typeorm';
 
 const mediaRoutes = Router();
 
@@ -70,8 +70,6 @@ mediaRoutes.get('/', async (req, res, next) => {
     case 'pending':
       statusFilter = MediaStatus.PENDING;
       break;
-    default:
-      statusFilter = undefined;
   }
 
   let sortFilter: FindOneOptions<Media>['order'] = {
@@ -90,12 +88,18 @@ mediaRoutes.get('/', async (req, res, next) => {
       };
   }
 
+  let whereClause: FindOneOptions<Media>['where'];
+  if (statusFilter || req.query.sort === 'mediaAdded') {
+    whereClause = {};
+    if (statusFilter) whereClause.status = statusFilter;
+    if (req.query.sort === 'mediaAdded')
+      whereClause.mediaAddedAt = Not(IsNull());
+  }
+
   try {
     const [media, mediaCount] = await mediaRepository.findAndCount({
       order: sortFilter,
-      where: statusFilter && {
-        status: statusFilter,
-      },
+      where: whereClause,
       take: pageSize,
       skip,
     });
@@ -134,7 +138,7 @@ mediaRoutes.post<
       return next({ status: 404, message: 'Media does not exist.' });
     }
 
-    const is4k = Boolean(req.body.is4k);
+    const is4k = String(req.body.is4k) === 'true';
 
     switch (req.params.status) {
       case 'available':
@@ -167,16 +171,16 @@ mediaRoutes.post<
             message: 'Only series can be set to be partially available',
           });
         }
-        media.status = MediaStatus.PARTIALLY_AVAILABLE;
+        media[is4k ? 'statusAlt' : 'status'] = MediaStatus.PARTIALLY_AVAILABLE;
         break;
       case 'processing':
-        media.status = MediaStatus.PROCESSING;
+        media[is4k ? 'statusAlt' : 'status'] = MediaStatus.PROCESSING;
         break;
       case 'pending':
-        media.status = MediaStatus.PENDING;
+        media[is4k ? 'statusAlt' : 'status'] = MediaStatus.PENDING;
         break;
       case 'unknown':
-        media.status = MediaStatus.UNKNOWN;
+        media[is4k ? 'statusAlt' : 'status'] = MediaStatus.UNKNOWN;
     }
 
     await mediaRepository.save(media);
@@ -196,15 +200,24 @@ mediaRoutes.delete(
         where: { id: Number(req.params.id) },
       });
 
-      await mediaRepository.remove(media);
+      if (media.status === MediaStatus.BLOCKLISTED) {
+        media.resetServiceData();
+        await mediaRepository.save(media);
+      } else {
+        await mediaRepository.remove(media);
+      }
 
       return res.status(204).send();
     } catch (e) {
-      logger.error('Something went wrong fetching media in delete request', {
+      if (e instanceof EntityNotFoundError) {
+        return res.status(204).send();
+      }
+      logger.error('Something went wrong deleting media', {
         label: 'Media',
+        mediaId: req.params.id,
         message: e.message,
       });
-      next({ status: 404, message: 'Media not found' });
+      next({ status: 500, message: 'Failed to delete media' });
     }
   }
 );
@@ -239,14 +252,15 @@ mediaRoutes.delete(
         );
       }
 
+      const specificServiceId = isAlt ? media.serviceIdAlt : media.serviceId;
       if (
-        media.serviceId &&
-        media.serviceId >= 0 &&
-        serviceSettings?.id !== media.serviceId
+        specificServiceId &&
+        specificServiceId >= 0 &&
+        serviceSettings?.id !== specificServiceId
       ) {
         if (isMovie) {
           serviceSettings = settings.radarr.find(
-            (radarr) => radarr.id === media.serviceId
+            (radarr) => radarr.id === specificServiceId
           );
         } else if (isBook) {
           serviceSettings = settings.readarr.find(
@@ -254,7 +268,7 @@ mediaRoutes.delete(
           );
         } else {
           serviceSettings = settings.sonarr.find(
-            (sonarr) => sonarr.id === media.serviceId
+            (sonarr) => sonarr.id === specificServiceId
           );
         }
       }
@@ -296,13 +310,10 @@ mediaRoutes.delete(
       }
 
       if (isMovie) {
-        await (service as RadarrAPI).removeMovie(
-          parseInt(
-            isAlt
-              ? (media.externalServiceSlugAlt as string)
-              : (media.externalServiceSlug as string)
-          )
-        );
+        if (!media.hasTmdbId()) {
+          throw new Error('TMDB ID is missing for this media!');
+        }
+        await (service as RadarrAPI).removeMovie(media.tmdbId);
       } else if (isBook) {
         if (!media.hasHcId()) {
           throw new Error('Hardcover ID is missing for this media!');
@@ -318,7 +329,7 @@ mediaRoutes.delete(
         if (!tvdbId) {
           throw new Error('TVDB ID not found');
         }
-        await (service as SonarrAPI).removeSerie(tvdbId);
+        await (service as SonarrAPI).removeSeries(tvdbId);
       }
 
       return res.status(204).send();
@@ -362,12 +373,12 @@ mediaRoutes.get<{ id: string }, MediaWatchDataResponse>(
       if (media.ratingKey) {
         const watchStats = await tautulli.getMediaWatchStats(media.ratingKey);
         const watchUsers = await tautulli.getMediaWatchUsers(media.ratingKey);
+        const plexIds = watchUsers.map((u) => u.user_id);
+        if (!plexIds.length) plexIds.push(-1);
 
         const users = await userRepository
           .createQueryBuilder('user')
-          .where('user.plexId IN (:...plexIds)', {
-            plexIds: watchUsers.map((u) => u.user_id),
-          })
+          .where('user.plexId IN (:...plexIds)', { plexIds })
           .getMany();
 
         const playCount =
@@ -394,12 +405,12 @@ mediaRoutes.get<{ id: string }, MediaWatchDataResponse>(
         const watchUsers4k = await tautulli.getMediaWatchUsers(
           media.ratingKeyAlt
         );
+        const plexIds4k = watchUsers4k.map((u) => u.user_id);
+        if (!plexIds4k.length) plexIds4k.push(-1);
 
         const users = await userRepository
           .createQueryBuilder('user')
-          .where('user.plexId IN (:...plexIds)', {
-            plexIds: watchUsers4k.map((u) => u.user_id),
-          })
+          .where('user.plexId IN (:...plexIds)', { plexIds: plexIds4k })
           .getMany();
 
         const playCount =
